requirements.txt
alembic==0.9.9
certifi==2018.4.16
chardet==3.0.4
click==6.7
Flask==1.0.2
Flask-Admin==1.5.1
Flask-API==1.0
Flask-Cors==3.0.4
flask-marshmallow==0.9.0
Flask-Migrate==2.1.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
future==0.16.0
idna==2.6
itsdangerous==0.24
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.0
marshmallow==2.15.3
marshmallow-sqlalchemy==0.13.2
MyApplication==0.1.0
numpy==1.14.3
pandas==0.23.0
passlib==1.7.1
patsy==0.5.0
psycopg2==2.7.4
psycopg2-binary==2.7.4
py-postgresql==1.2.1
PyMySQL==0.8.1
python-dateutil==2.7.3
python-editor==1.0.3
pytz==2018.4
PyYAML==3.12
requests==2.18.4
six==1.11.0
SQLAlchemy==1.2.7
SQLAlchemy-Utils==0.33.3
statsmodels==0.9.0
urllib3==1.22
Werkzeug==0.14.1
WTForms==2.1
